rename according to function anme
