+++ packages/render/src/edge/render.tsx
@@ -5,7 +5,7 @@ import type { Options } from '../shared/options';
 import { readStream } from '../shared/read-stream.browser';
 import { stripImagePreloadLinks } from '../shared/utils/strip-image-preload-links';
 import { toPlainText } from '../shared/utils/to-plain-text';
-import { unstableToPlainText } from '../shared/utils/to-plain-text-unstable';
+import { unstableToPlainText } from '../shared/utils/unstable-to-plain-text';
 import { importReactDom } from './import-react-dom';
 
 export const render = async (
