+++ packages/render/src/node/index.ts
@@ -1,5 +1,5 @@
 export * from '../shared/options';
 export * from '../shared/utils/pretty';
 export * from '../shared/utils/to-plain-text';
-export * from '../shared/utils/to-plain-text-unstable';
+export * from '../shared/utils/unstable-to-plain-text';
 export * from './render';
